# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "World UZ",
        "description": "Documentation for World Uz Company",
        "contact": {
            "email": "rustamovramziddin7@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Post a new data",
                "description": "Post new University  data",
                "operationId": "44212a9096e4b09358281e9ec8a0701d",
                "requestBody": {
                    "description": "Pass Auth   credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "password",
                                        "example": "123456789aawwee"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Post a new data",
                "description": "Post new user data",
                "operationId": "a718f172ff3ac464d723835815f8fb57",
                "requestBody": {
                    "description": "Pass Auth   credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username",
                                    "fullname",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "Doe"
                                    },
                                    "fullname": {
                                        "type": "text",
                                        "format": "",
                                        "example": "Surname First name"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "user@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "password",
                                        "example": "123456789aawwee"
                                    },
                                    "confirmation_password": {
                                        "type": "password",
                                        "format": "password",
                                        "example": "123456789aawwee"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Post a new data",
                "description": "Logout  data",
                "operationId": "79854151e609bd20032aa5aa4d5a35c0",
                "requestBody": {
                    "description": "Logout ",
                    "required": true
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "api": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    }
}